Feat: Serving dynamic public files
Servers can now serve all files in the public folder without explicitly defining them.
Upon starting the server, it performs a recursive call to find all files in the public folder and register them. When the files are registered, they are assigned to a controller, and when the controller receives a request for the file, it scans it and returns the file's type and content.
We create a new Utils function to recursively traverse the file and return its paths.

Currently, only HTML, CSS, and JS are supported.