feat(server option): Added SO_REUSEADDR for Development Env
The server option in development is set to forcefully use the address
without waiting for the timeout of the address.

It is done for live restart of the server on file change while
developing.
Note: This functionality may cause problem in production so it should
only be used while developing.