feat(client caching): Implemented ETag
Implemented ETag for client side caching.
The ETag uses the file size and last write to create a unique tag for
the data and send to the client.
Once the client again sends request to the server for the same resource
it checks the value of the tag and if same then 304 is sent.