feat(monitor): Improved working of monitor
Monitor now stops if the build is failed, before this change after
showing the error the server kept on running which made the error hard
to notice by the programmer.

Now user can choose to run a clean start of the server in monitor by
passing -c as one of the option.

Ubuntu setup how has to modes to run the development server, one the
normal modal and other the full clean mode.

This features improves the development experience.