feat:Major(Json): New Json Parser
Now the Json Parser uses two step to parse json, first it breaks the
json to tokens and check if any wrong token is present, then it parses
the tokens to build json object.

With the two step process now creating json now takes half the time, as
it doesnot need to iterate the token string twice and uses a linear and
recursive method in place of pure recursive.

The Lexer/Token step also helps to move complexity from the parser and
helps parser to focus on parsing.

The Json also has new intuitive methods which helps to easily access and
store data with simple name as set and get.

This is a Major change, and all Json use code was rewritted.