feat(json): Improved Json Parser by making scanner a generator
The scanner now doesn't return vector of tokens, but returns one token
at a time, this helps to remove the space used by the tokens and make it
fast.

After the changes the new Parser is taking 0.7sec in place of 1.5sec.

Known Issue:
Json Parser doesn't handle values greater than the range of long long,
it causes overflow in stod, for now if overflow occurs then the value is
set to 0.