feat(Major Change): Changed the Application to a library
Now Crystal can be used in any project as a library with the help of
CMake. Replaced the Makefile with CMakeLists.txt configurations.

Also added config class to configure the paths.
Also added scripts to build and run the project, this scripts are to
build run from the root directory.

The config path is also to be give from the root.

Note: Most of the functionality is been tested but some code me may
break because of the major shift